app/requirements.txt
click==7.1.2
cycler==0.10.0
decorator==4.4.2
Flask==1.1.2
imageio==2.9.0
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.1.2
networkx==2.4
numpy==1.18.0
opencv-python==4.1.2.30
Pillow==6.2.1
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
PyYAML==5.3.1
scikit-image==0.16.2
scipy==1.4.1
six==1.15.0
torch==1.3.0
torchvision==0.4.1
uWSGI==2.0.19.1
Werkzeug==1.0.1
